Protect the prefix tree flush process from inconsistent KV and indexer.
After a leader election, or a recovery, it is possible to have moved
backwards from the perspective of the "continued leader", in which case
the KV may contain an entry that is not known to the indexer yet.

Performing a flush in these circumstances would lead to the entry in the
KV to move backwards as well, where the latest entry would have a
smaller upper bound than its predecessor.

There were already asserts in batched_prefix_tree to protect against
this, but these only run in debug mode, and catch the problem much too
late.

We unfortunately still don't have a good way of testing these atypical
interleavings and interactions between the KV and the indexer.

Co-authored-by: Maik Riechert <maik.riechert@microsoft.com>