Update to CCF 3.0.
The most intrusive change necessary comes from CCF introducing
a notion of "platform" to distinguish between SGX and Virtual (and SNP,
although we don't yet support that).

A lot of places in build scripts and CI configuration which used a
"release" vs "virtual" enclave type use the now more obvious "sgx" vs
virtual platform. Additionally, CCF artifacts are now split by platform,
meaning we need to use different Docker images for the each platform.

Because CCF has stopped providing host libraries in their SGX
distribution, we cannot build and run unit tests as part of the SGX
pipeline anymore. This is mostly fine since the code behaves the same in
both builds, with the exception of the DID web resolution, which on SGX
requires the result to be attested. This code wasn't ever unit-tested so
it isn't a great loss (it isn't even clear whether this code would build
on the host, due to its dependency on OE libraries).

CCF now includes its own copy of QCBOR and t_cose, meaning we can remove
them from our build process.

It also adds support for COSE signatures for governance actions as an
alternative to HTTP signing, we which may eventually want to take
advantage. This is left as future work.