Detect the sgx_prv group ID in CI automatically.
During CI, we need to create the sgx_prv group inside the Docker
container and add the current user to it.

We used to hardcode 119 as the group ID, but this value has changed on
recent build machines.

Instead we can use stat to examine the /dev/sgx/provision file using
stat and obtain the group ID from its ownership.