Add an isolated_test marker for functional tests.
If a test class is annotated with the marker, a separate cchost process
is started for this test class. This provides isolation between tests,
which is useful either if the test is deemed risky and can leave the
service in an irrecoverably bad state (eg. by modifying the
constitution), or if the test depends on having an empty ledger at
start.

Note that when using an external service, it is impossible to provide
isolation and even tests marked with isolated_test will use the shared
instance. The marker can be combined with the needs_cchost marker to
skip the test in these cases where isolation cannot be guaranteed.