Return a nice error when no prefix tree has been committed yet.
If no prefix tree commit has been indexed yet, the field in the indexing
strategy that stores the seqno is 0. Previously the frontend would get
this sequence number and attempt to perform a historical query on it,
which led to CCF throwing an exception.

Instead we detect that case and return a nullopt from the indexing
strategy, which gets turned into a 404 to the client.

This was only an issue for the top-level `/app/prefix_tree` endpoint.
On the `/app/prefix_tree/{issuer}/{feed}` endpoint, looking up a receipt
in an empty/missing prefix tree always fails anyway, and was already
returning a 404 error.

A test for this is added, but is marked as XFAIL, as it cannot currently
be executed reliably since this only happens on an empty service. Once
the test framework supports starting isolated services for individual
test cases, we can remove the XFAIL marker.

In the future we may want to automatically issue a flush for the empty
prefix tree on service startup. This won't eliminate the problem
completely since there will be a window of time between the automatic
initial flush and it getting indexed where no prefix tree receipt is
available, but we would consider that a transient IndexingInProgress
error instead.