Add iss and kid fields to receipts.
We're headed towards using a DID for the server's identity. This will
require a few parts, but the first step is to add iss and kid fields to
the receipt's protected headers.

The iss value is be copied from an optional field in the configuration
file. If the field is missing, we omit the iss and kid altogether. The
kid should represent the current service's certificate/public key and
will change on disaster recovery. For now, we use the same value we use
for service_identity, ie. the hash of the DER certificate, but may
revisit this in the future (eg. a hash of just the public key).

The old service_identity field does not get removed just yet, so as to
not break tooling that uses it while we are still transitioning to DIDs.
In particular, pyscitt and our functional tests still depend on it.