Add a flag to tests to start cchost. (#21)
We replace sandbox.sh with our own Python implementation, in
`test/infra/cchost.py`.

The functional tests can be called with a `--start-cchost`, in which
case the test infrastructure will launch a start and configure a new
cchost process for the duration of the test session. This provides
tighter and more powerful integration with the tests and opens up the
door for new possibilities, such as starting an isolated cchost process
for certain tests or performing a disaster recovery on the service. It
is also friendlier to use, as a single command and console output
manages everything.

It is still possible to run the tests against an external service, which
is necessary to run the Docker based test jobs. While most test support
this scenario, in the future individual tests may be annotated with
`needs_cchost` and will be skipped if using an external service.

The `cchost.py` file may also be executed as a standalone script, which
is useful for local development. The `run.sh` script has been updated to
use this script instead of `sandbox.sh`.

Fixes #6.

Co-authored-by: Maik Riechert <maik.riechert@microsoft.com>