Add integration test for non-canonically encoded protected headers.
The COSE spec does not impose any encoding requirements for the contents
of the protected headers, but instead requires that these are wrapped in
a bstr and never re-encoded by any party processing the message.

This adds some sanity checks, mostly focusing on these points:
- The ledger accepts claims with a non-canonical encoding
- When embedding a receipt, the ledger does not affect the encoding of
  the claim.
- pyscitt can verify the receipts for these claims.

The last point is actually broken for now, due to a bug in the current
pycose release. This has been fixed upstream, but hasn't been released
yet. The relevant test is marked as XFAIL, until pycose is updated.