Change the signature in receipts from ASN1/DER to IEEE encoding.
So far, we has been encoding the ECDSA signatures in receipts using the
ASN1/DER encoding that is commonly found in the x509 ecosystem, and
which OpenSSL and CCF use. This encoding prefixes each of the two
components with a tag and length, and groups the two in a SEQUENCE.

By contrast, the IEEE encoding is used by the COSE and JOSE ecosystems
already. It is merely the two integer values padded to the curve size
and concatenated together.

This is a breaking change, as old receipts cannot be verified by newer
clients, and vice-versa. As the signature encoding happens on-demand
when receipts are requested, if existing ledger is updated all receipts
it issues going forward will use the new format, including for claims
that were registered prior to the upgrade.

Fixes #49