Include all past service identities in the DID endpoint. (#85)
This replaces the direct KV lookup with an indexing strategy that
collects and keeps in memory all past identities. This allows the
endpoint to return a full DID document with all these identities,
without the need for historical queries and retries.

This PR adds support for disaster recovery to our test infrastructure,
by restarting the cchost process. Because cchost port numbers are
randomly assigned by the operating system, there would not be a stable
address on which to contact the service. This is problematic both
because we'd have to update `Client` instances to point to the new
service, and because the service's address is used as its did:web long
term identifier. To circumvent both problems, a Python TCP proxy is
used, listening on a unique port for the entire duration of a test, and
dynamically switching the destination of connections as the cchost
process is restarted.

Fixes #53