Fix a use after free bug in LRU.
The LRU cache picks the last item from its queue, removes it from the
list and invokes a custom callback. However, removing the entry from the
list deallocates it, which leads to a use-after-free when invoking the
callback. Instead it should first invoke the callback, and only then
remove it from the list using `pop_back()`.

This was found in CI job running ASan. I think this just happened to be
the first time an ASan build reached the threshold where the cache is
full. It can easily be reproduced this on a local ASan build, by
submitting hundreds of claims and fething receipts for all of them.

Added a new property-based test for the LRU cache to check for basic
correctness of it, and which would have caught this bug.