Add a `service_error` helper to the test infrastucture.
It is very common for our integration tests to check for errors returned
by the server, and this is generally accomplished by using pytest's
`raises` context manager, eg:

    with pytest.raises(ServiceError, match="expected message"):
        client.do_something()

Because the expected message is often long, this regularly doesn't fit
on a single line and needs to be broken up and looks messy and hard to
read. By having a short hand for this we can avoid the line breaks most
of the time.

The short-hand version already existed in test_encoding, but is now
moved to a shared file and used consistently.