Add integration tests for the DID resolver.
Some of the resolver's functionality depends on the host's system time.
In order to test these parts, the test infrastucture now uses the
libfaketime utility to intercept library calls to time functions, and
allow us to introduce an artificial time shift.

A lot of the resolver's error handling cannot yet be tested, because
errors aren't surfaced through the API. When fixing #108, we'll be able
to add a lot more tests.