Implement the new REST API (#99)
This introduces the concept of a long-running operation to the API. When
submitting a claim, the ledger returns a temporary operation ID. Clients
must poll an endpoint to check the status of the operation.

Eventually, the operation will complete and the status endpoint will
return long-term entry ID, which the client can use to retrieve a
receipt.

This schema can be used to model both the DID-web resolution phase and
the CCF consensus process.

If submitting a claim requires a DID resolution, the initial POST to the
/claims endpoint creates a new operation by saving the claim to a
special KV, and triggers the external process to fetch the DID document.
When the external process invokes the operation callback with the
resolution payload, the ledger uses a historical to get the original
claim back. Once both the claim and DID document are available, the
existing registration process takes place and a new row is added to the
KV to mark the operation as complete.

After a claim has been written the the KV, either synchronously in the
original POST request or asynchronously in the resolution callback, the
operation is still considered as pending, until the transaction is
globally committed and a receipt is available. At this point only will
the operation be considered completed.

This new API has multiple benefits:

- It surfaces DID resolution errors to the end-user. Previously, clients
would only ever receive "resolution in progress" messages, but never
any hard errors with informative description. Now the operation can be
marked as failed by writing details of the error to the KV, and the
error message is returned when a client polls the operation status.
- It is a more intuitive API and does not require clients to retry POST
requests anymore. The only request that needs polling is a GET to
check the operation status, and one day this may even be replaced by a
long-polling request.
- It unifies different delays, such as resolution and consensus, under a
single interface, abstracting implementation details from clients.

Fixes #108