Delay the final tracing message until after the local commit handler. (#117)
When a local commit handler is installed, it can modify the response's
status code after the main endpoint handler has returned. When this
happened, the log message printed by the tracing adapter would be
inaccurate, as it wouldn't take into account the changes made by the
commit handler into account.

The printing of the log message is instead delayed until the very end of
the local commit message, when we know the application won't change the
response anymore. Note that if the main endpoint handler returned a
non-2xx status, CCF never calls the commit handler, so we have to print
the log message early in this case.

Fixes #114