Update to CCF 3.0.9.
The new version now captures the output of host subprocesses and prints
it onto the service's logs, which makes debugging DID web resolution
much easier.

The fetch-did-web-doc.sh script, whose only purpose was to redirect the
output to a file on disk, is not necessary anymore and has been removed.
The service now invokes the Python script directly. Additionally, the
path to the script is now configurable at compile time.

Fixes #119 and #77.