Switch the operations callback to use an indexing strategy. (#140)
The callback endpoint when completing an operation needs to access the KV to check the integrity of the operation context that is being passed by the caller.

We previously used as historical query to look this up, but this almost always ends up returning a Service Unavailable error while the historical transaction is being fetched, and the callback needs to be retried after a short delay.

This replaces the historical query by caching the expected digest in the existing operations indexing strategy. This increases memory usage a little, but given that the indexing strategy is periodically purged, this increase should be bounded. If the operation completes too quickly, there is still a chance the strategy won't have indexed the original transaction yet, and the existing retry mechanism will be used.